Update poissondistribution.tex
If line 388 is not necessary to be changed, then maybe this line should also be '\lambda+\mu' to be consistent.  Since they are asking the same question. But I think maybe it's better to modify line 388 because in previous context P(\lambda t) is used to describe Poisson process? I'm not sure which way is preferred.